Initial commit
This is the barebones structure of the project. It currently generates a
random league of 30 teams with 15 players each and basically has two
point guards with the same name play 1-on-1 against each other for four
quarters. Includes shooting and rebounding, to an extent, as well as a
shot clock and game clock. Shot results and score are updated after
every shot to System.out. Very happy with progress in first couple
sittings, clear list of features to add so I'll be doing that in the
coming weeks.