Improved simulation, persistent stats, and more
Stats are now recorded for the season, could do some interesting things
there. The simulation seems better after some adjustments to the shoot()
function, taking into account defensive ratings and some tweaks to
randomization. Finally fixed the bug where endGame() was called a ton of
times. Still need to find a way to "save the game," by recording the
league info and stats somehow and letting it be loaded back in.