82 game season... reveals bugs
Something is wrong with the way I'm storing players either during games
or between games. Definitely need to tinker with this. On the bright
side, it seems that bugs aside, it should be easy to turn my simple game
simulator into a "season simulator" with a calendar, etc., which is the
next step.